refactor: move some test utils to e2e folder
