+++ packages/e2e-performance/src/utils/test/sample-gfxinfo-output.txt
